chore(Network proxy): move connection wrappers to `requests` package
