+++ extensions/shared-youtube/library/src/main/java/app/morphe/extension/shared/requests/proxy/DelegatingHttpURLConnection.java
@@ -5,7 +5,7 @@
  * See the included NOTICE file for GPLv3 Section 7 terms that apply to this code.
  */
 
-package app.morphe.extension.shared.patches;
+package app.morphe.extension.shared.requests.proxy;
 
 import java.io.IOException;
 import java.io.InputStream;
